Configurable apiserver resources (limits/requests)
In some cases, it is possible for the apiserver to leak memory (see
https://github.com/kubernetes/kubernetes/issues/65732). When the node
runs out of memory, the master can struggle to recover and cause
slowdowns and API outages. A mitigation to this is to set explicit
memory limits to pre-emptively terminate the apiserver and cycle memory
usage.

Without this change, it is necessary to SSH into each master, manually
update the static pod manifest, and restart kubelet. This allows us to
configure apiserver resources definitively.